Add foldable panes into main webpages
This adapts and extends the Boostrap accordion example to have
foldable sections in the main webpages of the study groups. This
could be used to structure and organize content into subpages. The
main files and structure:

_data/accordion.yml:
Each mapping (e.g., subspace) represents a main page (e.g., the
main page for the multivariate subspace group). For each mapping,
one can add one block per subpage (e.g., the 'resources' page for
the subspace group). There are three fields on each block: 'title',
'url', and 'id'. These fields are to be used in the accordion
template to generate the foldable sections.

_includes/accordion.html:
This contains the Boostrap code and Liquid/Jekyll code to build
the foldable panes using the data from accordion.yml and attributes
from the variables page and site.

In the page where the template accordion.html is used, one must
define a subfolder variable in the front matter indicating which
mapping in accordion.yml will be used. The subspages has to be
named as <id>.md, where <id> is the corresponding id field in
the accordion.yml file. Also, this subpages have to be stored in
a subfolder called <subfolder>, inside the _study-groups folder.

In the foldable panes, a preview or excerpt of each subpage will
be presented. By default, the first paragraph is displayed, but
one can control how much to show by adding <!--more--> at the
point in the subpage where one wants to stop the preview.

This commit includes an example about how to use the accordion
template for the multivariate blind subspace methods study group,
but the template is general enough to be used by other study groups.